+++ src/main/java/community/whatever/onembackendjava/service/AdminUrlShortenService.java
@@ -1,5 +1,6 @@
-package community.whatever.onembackendjava;
+package community.whatever.onembackendjava.service;
 
+import community.whatever.onembackendjava.UrlMappingManager;
 import community.whatever.onembackendjava.dto.HashMapResponse;
 import community.whatever.onembackendjava.dto.PostShortenUrlsRequest;
 import lombok.RequiredArgsConstructor;
